Added history action.
A request for a wiki page with the history action returns a basic page
of the last 5 (hard-coded for now) commits for that file.  The next
step is to show all commits divided into multiple pages (say, 10
commits a page), and make each entry link to the that version of the
page, which will require an action for accessing specific commits.